quell size_t to int warning
svn:revision:72729
svn:branch:trunk
svn:account:brlcad

+++ misc/tools/perplex/perplex_template.c
@@ -284,7 +284,7 @@ buf_strnappend(struct Buf *buf, const char *str, int n)
 struct Buf*
 buf_strappend(struct Buf *buf, const char *str)
 {
-    return buf_strnappend(buf, str, strlen(str));
+  return buf_strnappend(buf, str, (int)strlen(str));
 }
 
 /* appends "#define str def\n" */
